Rename tower into tower_elasto
