+++ src/elasto/tower_elasto.cpp
@@ -1,4 +1,4 @@
-#include "tower.h"
+#include "tower_elasto.h"
 #include "../utils.h"
 
 void Tower::build(std::shared_ptr<ChMesh> mesh) {
